changed user_rating to intger and the input to range from 0 - 5

+++ products/models.py
@@ -132,8 +132,7 @@ class UserReview(models.Model):
     user = models.ForeignKey(User,
                              on_delete=models.CASCADE,
                              related_name="userreviews")
-    user_rating = models.DecimalField(
-        max_digits=6, decimal_places=2, null=True, blank=True)
+    user_rating = models.IntegerField(null=True, blank=True)
 
     def __str__(self):
         return f'''
